[Footnote] Move LicenseBox into dedicated template
