+++ pages/blog/[...slug].vue
@@ -62,55 +62,7 @@
                           </a>
                       </div>
                     </div>
-                    <div class="license-box my-3">
-                      <div class="license-title">
-                        <div>{{doc.title}}</div>
-                        <div>https://next.inori.moe/{{doc._path}}</div>
-                      </div>
-                      <div class="license-meta">
-                        <div class="license-meta-item">
-                          <div>Author</div>
-                          <div>inori</div>
-                        </div>
-                        <div class="license-meta-item license-meta-date">
-                          <div>Published</div>
-                          <div>{{ renderDate(doc.date) }}</div>
-                        </div>
-                        <div class="license-meta-item">
-                          <div>License</div>
-                          <div>
-                            <a class="print-no-link" target="_blank" href="https://creativecommons.org/licenses/by-nc-nd/4.0/">
-                              <span class="hint--top hint--rounded" aria-label="BY - 署名">
-                                <i class="iconfont icon-by">
-                                  <Icon name="cib:creative-commons-by" />
-                                </i>
-                              </span>
-                            </a>
-                            <a class="print-no-link" target="_blank" href="https://creativecommons.org/licenses/by-nc-nd/4.0/">
-                              <span class="hint--top hint--rounded" aria-label="NC - 非商业使用">
-                                <i class="iconfont icon-nc">
-                                  <Icon name="cib:creative-commons-nc" />
-                                </i>
-                              </span>
-                            </a>
-                            <a class="print-no-link" target="_blank" href="https://creativecommons.org/licenses/by-nc-nd/4.0/">
-                              <span class="hint--top hint--rounded" aria-label="ND - 禁止演绎">
-                                <i class="iconfont icon-nd">
-                                  <Icon name="cib:creative-commons-nd" />
-                                </i>
-                              </span>
-                            </a>
-                          </div>
-                        </div>
-                      </div>
-                      <div id="cc-large-icon">
-                        <i>
-                          <Icon name="cib:creative-commons"
-                                size="200px"
-                          />
-                        </i>
-                      </div>
-                    </div>
+                    <LicenseBox :title="doc.title" :path="doc._path" :date="doc.date" />
                     <!-- TODO: prev/next cannot be currently implemented -->
                   </div>
                   <!-- TODO: implement comment section -->
@@ -135,11 +87,8 @@ import Footer from "~/views/Footer.vue";
 import {onMounted} from "#imports";
 import PartialCommonHeader from "~/views/header/PartialCommonHeader.vue";
 import NavBar from "~/views/header/NavBar.vue";
+import LicenseBox from "~/views/LicenseBox.vue";
 
-const renderDate = (date: string) => {
-  const dateParts = date.split("T")[0].split('-')
-  return `${dateParts[0]}年${dateParts[1]}月${dateParts[2]}日`
-}
 
 let scrollHeight = ref(0)
